+++ Classes/FSATextureManager.mm
@@ -188,8 +188,8 @@ static FSATextureManager* fsaTextureManager;
     
     CGContextSetTextDrawingMode(context, kCGTextFillStroke);
     
-    float components[2] = { 1, 1 };
-    float components2[2] = { .26666667, 1 };
+    CGFloat components[2] = { 1, 1 };
+    CGFloat components2[2] = { .26666667, 1 };
     CGColorRef color = CGColorCreate(colorSpace, components);
     CGColorRef color2 = CGColorCreate(colorSpace, components2);
